Cleaning up the Gruntfile into the grunttasks dir
