+++ grunttasks/build.js
@@ -1,0 +1,28 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+module.exports = function (grunt) {
+  'use strict';
+
+  grunt.registerTask('build', [
+    'lint',
+    'clean:dist',
+    'useminPrepare',
+    'selectconfig:dist',
+    'l10n-create-json',
+    'requirejs',
+    'concurrent:dist',
+    'autoprefixer',
+    'concat',
+    'cssmin',
+    'copy:dist',
+    // modernizer must come after copy or else the custom
+    // modernizr is overwritten with the dev version.
+    'modernizr',
+    // uglify overwrites the files in the dist directory.
+    'uglify',
+    'rev',
+    'usemin'
+  ]);
+};
